+++ al-khaser/Anti VM/Qemu.h
@@ -6,3 +6,5 @@
 
 VOID qemu_reg_key_value();
 VOID qemu_processes();
+BOOL qemu_firmware_ACPI();
+BOOL qemu_firmware_SMBIOS();