+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #### 0.74:
-- Added qemu process check (qemu-ga.exe) thanks to kaganisildak.
-
+- Added qemu process check (qemu-ga.exe) thanks to @kaganisildak.
+- Added checks for system firmware tables (SMBIOS and ACPI for QEMU).
 
 #### 0.73
 - Bug fix: GetSystemFirmwareTable should take `resultBufferSize` as an argument for the second call.
@@ -21,7 +21,7 @@
 - Bug fix: array with duplicate strings in process_tools check thanks to @stxletto.
 - Bug fix: ascii_to_wide_str() wrong argument thanks to @stxletto.
 
-### 0.71
+#### 0.71
 - New: Add kernel debugger check using the KUSER_SHARED_DATA struct.
 - New: Add kernel debugger check using NtQuerySystemInformation with SystemKernelDebuggerInformation.
 - New: Added process job anti-debug check.